+++ src/entities/multi-key.js
