+++ src/entities/player.js
